refactor: improve error handling and refactor output silencing utilities
