+++ cmd/start.go
@@ -46,7 +46,7 @@ func start() error {
 
 	ready, err := containerEngineProvisioner.CheckReady()
 	if err != nil || !ready {
-		return fmt.Errorf("container engine '%s' is not ready: %v", ksailConfig.Spec.ContainerEngine, err)
+		return fmt.Errorf("container engine '%s' is not ready: %w", ksailConfig.Spec.ContainerEngine, err)
 	}
 
 	fmt.Printf("✔ '%s' is ready\n", ksailConfig.Spec.ContainerEngine)
